requirements.txt
cachetools==5.3.0
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.92.0
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
h11==0.14.0
httpcore==0.16.3
httplib2==0.21.0
httptools==0.5.0
httpx==0.23.3
itsdangerous==2.1.2
orjson==3.8.7
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.5
pyparsing==3.0.9
python-dotenv==1.0.0
python-multipart==0.0.6
rfc3986==1.5.0
rsa==4.9
starlette==0.25.0
ujson==5.7.0
uritemplate==4.1.1
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
